+++ src/main/pipe-server.ts
@@ -40,7 +40,7 @@ export class PipeServer extends EventEmitter {
   private pipePath: string;
   private authToken: string;
 
-  constructor(pipePath = '\\\\.\\pipe\\wmux', authToken = '') {
+  constructor(pipePath = '\\\\.\\pipe\\winagent-terminal', authToken = '') {
     super();
     this.pipePath = pipePath;
     this.authToken = authToken;